split inventories into files
