+++ BetterInventory.csproj
@@ -5,6 +5,7 @@
     <nullable>enable</nullable>
   </PropertyGroup>
   <ItemGroup>
-    <ProjectReference Include="..\SpikysLib\SpikysLib.csproj" />
+    <Reference Include="../../ModAssemblies/SpikysLib_v1.0.dll" />
+    <!-- <ProjectReference Include="../SpikysLib/SpikysLib.csproj" /> -->
   </ItemGroup>
 </Project>
