+++ ItemSearch/BetterNPC.cs
@@ -4,7 +4,7 @@ using Terraria.ModLoader;
 
 namespace BetterInventory.ItemSearch;
 
-public class BetterNPC : GlobalNPC {
+public sealed class BetterNPC : GlobalNPC {
     public override void OnChatButtonClicked(NPC npc, bool firstButton) {
         if (!Configs.BetterGuide.Enabled || npc.type != NPCID.Guide) return;
         Main.InGuideCraftMenu = true;
