+++ ModSubInventory.cs
